add remove command to cli to prune all infrastructure
