+++ .gitignore
@@ -1,0 +1,17 @@
+# Terraform
+*.tfstate
+*.tfstate.backup
+*.tfplan
+.terraform/
+terraform/http/lib/node_modules
+terraform/https/lib/node_modules
+terraform/http/.terraform
+terraform/https/.terraform
+terraform/http/.tfstate
+terraform/http/.tfstate.backup
+terraform/https/.tfstate
+terraform/https/.tfstate.backup
+
+.DS_Store
+lib/.DS_Store
+*.pem
