--- terraform/.gitignore
@@ -1,17 +1,0 @@
-# Terraform
-*.tfstate
-*.tfstate.backup
-*.tfplan
-.terraform/
-http/lib/node_modules
-https/lib/node_modules
-http/.terraform
-https/.terraform
-http/.tfstate
-http/.tfstate.backup
-https/.tfstate
-https/.tfstate.backup
-
-.DS_Store
-lib/.DS_Store
-*.pem
